Move posting the client-to-user messages on the assistant queue into `client_to_user` function.
